+++ src/dataManagers/auth.js
