+++ src/dataManagers/seasonLogs.js
